requirements.txt
Flask==2.3.2
gunicorn==21.2.0
faiss-cpu
sentence-transformers
transformers
torch
scikit-learn
